+++ NEWS.md
@@ -4,6 +4,9 @@
 - Add featureData slot to Chromatograms class and add mz, precursorMz and 
   productMz methods for Chromatograms (issue
   [#289](https://github.com/lgatto/MSnbase/issues/289)) <2017-12-18 Mon>.
+- The MSnSetList class has a new `featureData` slot, accessible with
+  `fData` to store metadata for the individial MSnSets of the
+  list. `MSnSetList` also not has an `sapply` method. <2018-01-10 Wed>
 
 ## Changes in version 2.5.3
 - MSnSet to/from SummarizedExperiment coercion (contributed by Arne